+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Hydro Raindrop SDK PHP
 
+## 1.2.0
+
+* Lower PHP requirement to `7.0`.
+* [BC] Refactor `ApiAccessToken` method `getExpiresIn` to `getExpiresAt`.
+* [BC] Refactor `TokenStorage` method `getAccessToken` to not be nullable anymore, 
+  instead the exception `UnableToAcquireAccessToken` will be thrown when an access token 
+  cannot be acquired from the storage. 
+
 ## 1.1.2
 
 * Add proper `@throws` PhpDoc for some functions.